+++ src/type_object.rs
@@ -1,13 +1,13 @@
 // Copyright (c) 2017-present PyO3 Project and Contributors
 //! Python type object information
 
+use crate::once_cell::GILOnceCell;
 use crate::pyclass::{initialize_type_object, PyClass};
 use crate::pyclass_init::PyObjectInit;
 use crate::types::{PyAny, PyType};
-use crate::{ffi, AsPyPointer, Python};
-use std::cell::UnsafeCell;
-use std::ptr::NonNull;
-use std::sync::atomic::{AtomicBool, Ordering};
+use crate::{ffi, AsPyPointer, PyNativeType, Python};
+use parking_lot::{const_mutex, Mutex};
+use std::thread::{self, ThreadId};
 
 /// `T: PyLayout<U>` represents that `T` is a concrete representaion of `U` in Python heap.
 /// E.g., `PyCell` is a concrete representaion of all `pyclass`es, and `ffi::PyObject`
@@ -101,18 +101,21 @@ pub unsafe trait PyTypeInfo: Sized {
     type AsRefTarget: crate::PyNativeType;
 
     /// PyTypeObject instance for this type.
-    fn type_object() -> &'static ffi::PyTypeObject;
+    fn type_object_raw(py: Python) -> &'static ffi::PyTypeObject;
 
     /// Check if `*mut ffi::PyObject` is instance of this type
     fn is_instance(object: &PyAny) -> bool {
         unsafe {
-            ffi::PyObject_TypeCheck(object.as_ptr(), Self::type_object() as *const _ as _) != 0
+            ffi::PyObject_TypeCheck(
+                object.as_ptr(),
+                Self::type_object(object.py()) as *const _ as _,
+            ) != 0
         }
     }
 
     /// Check if `*mut ffi::PyObject` is exact instance of this type
     fn is_exact_instance(object: &PyAny) -> bool {
-        unsafe { (*object.as_ptr()).ob_type == Self::type_object() as *const _ as _ }
+        unsafe { (*object.as_ptr()).ob_type == Self::type_object(object.py()) as *const _ as _ }
     }
 }
 
@@ -132,78 +135,63 @@ where
     T: PyTypeInfo,
 {
     fn type_object(py: Python) -> &PyType {
-        unsafe { py.from_borrowed_ptr(<Self as PyTypeInfo>::type_object() as *const _ as _) }
+        unsafe { py.from_borrowed_ptr(Self::type_object_raw(py) as *const _ as _) }
     }
 }
-
-/// Lazy type object for Exceptions
-#[doc(hidden)]
-pub struct LazyHeapType {
-    value: UnsafeCell<Option<NonNull<ffi::PyTypeObject>>>,
-    initialized: AtomicBool,
-}
-
-impl LazyHeapType {
-    pub const fn new() -> Self {
-        LazyHeapType {
-            value: UnsafeCell::new(None),
-            initialized: AtomicBool::new(false),
-        }
-    }
-
-    pub fn get_or_init<F>(&self, constructor: F) -> NonNull<ffi::PyTypeObject>
-    where
-        F: Fn(Python) -> NonNull<ffi::PyTypeObject>,
-    {
-        if !self
-            .initialized
-            .compare_and_swap(false, true, Ordering::Acquire)
-        {
-            // We have to get the GIL before setting the value to the global!!!
-            let gil = Python::acquire_gil();
-            unsafe {
-                *self.value.get() = Some(constructor(gil.python()));
-            }
-        }
-        unsafe { (*self.value.get()).unwrap() }
-    }
-}
-
-// This is necessary for making static `LazyHeapType`s
-//
-// Type objects are shared between threads by the Python interpreter anyway, so it is no worse
-// to allow sharing on the Rust side too.
-unsafe impl Sync for LazyHeapType {}
 
 /// Lazy type object for PyClass
 #[doc(hidden)]
 pub struct LazyStaticType {
-    value: UnsafeCell<ffi::PyTypeObject>,
-    initialized: AtomicBool,
+    // Boxed because Python expects the type object to have a stable address.
+    value: GILOnceCell<Box<ffi::PyTypeObject>>,
+    // Threads which have begun initialization. Used for reentrant initialization detection.
+    initializing_threads: Mutex<Vec<ThreadId>>,
 }
 
 impl LazyStaticType {
     pub const fn new() -> Self {
         LazyStaticType {
-            value: UnsafeCell::new(ffi::PyTypeObject_INIT),
-            initialized: AtomicBool::new(false),
+            value: GILOnceCell::new(),
+            initializing_threads: const_mutex(Vec::new()),
         }
     }
 
-    pub fn get_or_init<T: PyClass>(&self) -> &ffi::PyTypeObject {
-        if !self
-            .initialized
-            .compare_and_swap(false, true, Ordering::Acquire)
-        {
-            let gil = Python::acquire_gil();
-            let py = gil.python();
-            initialize_type_object::<T>(py, T::MODULE, unsafe { &mut *self.value.get() })
-                .unwrap_or_else(|e| {
-                    e.print(py);
-                    panic!("An error occurred while initializing class {}", T::NAME)
-                });
-        }
-        unsafe { &*self.value.get() }
+    pub fn get_or_init<T: PyClass>(&self, py: Python) -> &ffi::PyTypeObject {
+        self.value
+            .get_or_init(py, || {
+                {
+                    // Code evaluated at class init time, such as class attributes, might lead to
+                    // recursive initalization of the type object if the class attribute is the same
+                    // type as the class.
+                    //
+                    // That could lead to all sorts of unsafety such as using incomplete type objects
+                    // to initialize class instances, so recursive initialization is prevented.
+                    let thread_id = thread::current().id();
+                    let mut threads = self.initializing_threads.lock();
+                    if threads.contains(&thread_id) {
+                        panic!("Recursive initialization of type_object for {}", T::NAME);
+                    } else {
+                        threads.push(thread_id)
+                    }
+                }
+
+                // Okay, not recursive initialization - can proceed safely.
+                let mut type_object = Box::new(ffi::PyTypeObject_INIT);
+
+                initialize_type_object::<T>(py, T::MODULE, type_object.as_mut()).unwrap_or_else(
+                    |e| {
+                        e.print(py);
+                        panic!("An error occurred while initializing class {}", T::NAME)
+                    },
+                );
+
+                // Initialization successfully complete, can clear the thread list.
+                // (No futher calls to get_or_init() will try to init, on any thread.)
+                *self.initializing_threads.lock() = Vec::new();
+
+                type_object
+            })
+            .as_ref()
     }
 }